Fix all clippy warnings across workspace
Address clippy lints to pass `cargo clippy -- -D warnings`:

- Replace manual div_ceil implementations with `.div_ceil()` method
- Collapse nested if statements where appropriate
- Convert needless range loops to iterator patterns
- Remove identity operations (e.g., `base + 0` → `base`)
- Add allow attributes for intentional patterns:
  - `#[allow(clippy::too_many_arguments)]` for spread functions
  - `#[allow(clippy::needless_range_loop)]` where indices used for bit ops
  - `#[allow(clippy::identity_op)]` for formula clarity in tests
  - `#[allow(clippy::unit_arg)]` in benchmarks using black_box
  - `#[allow(dead_code)]` for ticks field (bare-metal platforms)
  - `#[allow(unused_imports)]` for WIP circuit-bench modules
- Fix field_reassign_with_default using struct init syntax
- Remove unnecessary drop() calls on non-Drop types
- Use strip_prefix() instead of manual string slicing
- Replace redundant closures with function references

+++ prav-fb-bench/src/verification.rs
@@ -75,17 +75,17 @@ pub fn verify_fb_corrections(
     let mut syndrome_change = vec![false; num_nodes];
 
     for &edge_idx in edge_corrections {
-        if (edge_idx as usize) < edges.len() {
-            let (u, v, _) = edges[edge_idx as usize];
+        if edge_idx < edges.len() {
+            let (u, v, _) = edges[edge_idx];
 
             // Toggle at u if it's a real node (not boundary)
-            if (u as usize) < num_nodes {
-                syndrome_change[u as usize] ^= true;
+            if u < num_nodes {
+                syndrome_change[u] ^= true;
             }
 
             // Toggle at v if it's a real node (not boundary)
-            if (v as usize) < num_nodes {
-                syndrome_change[v as usize] ^= true;
+            if v < num_nodes {
+                syndrome_change[v] ^= true;
             }
         }
     }
@@ -93,8 +93,8 @@ pub fn verify_fb_corrections(
     // Build original syndrome as a set
     let mut original_syndrome = vec![false; num_nodes];
     for &d in defects {
-        if (d as usize) < num_nodes {
-            original_syndrome[d as usize] = true;
+        if d < num_nodes {
+            original_syndrome[d] = true;
         }
     }
